+++ CHANGELOG.md
@@ -3040,3 +3040,4 @@
 - [2026-07-22 14:42] perf(test): memoize expensive computations in test
 - [2026-07-22 14:49] style(test): harmonize light and dark theme palette for test
 - [2026-07-22 17:20] feat(auth): add bi-directional event bus handler for auth
+- [2026-07-22 17:38] perf(state): apply RepaintBoundary on animated state widgets